Rename CodeGenerator module to Console
