+++ src/Console/Infrastructure/Command/MakeFileCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Gacela\CodeGenerator\Infrastructure\Command;
+namespace Gacela\Console\Infrastructure\Command;
 
-use Gacela\CodeGenerator\CodeGeneratorFacade;
-use Gacela\CodeGenerator\Domain\FilenameSanitizer\FilenameSanitizer;
+use Gacela\Console\ConsoleFacade;
+use Gacela\Console\Domain\FilenameSanitizer\FilenameSanitizer;
 use Gacela\Framework\DocBlockResolverAwareTrait;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -14,39 +14,44 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @method CodeGeneratorFacade getFacade()
+ * @method ConsoleFacade getFacade()
  */
-final class MakeModuleCommand extends Command
+final class MakeFileCommand extends Command
 {
     use DocBlockResolverAwareTrait;
 
     protected function configure(): void
     {
-        $this->setName('make:module')
-            ->setDescription('Generate a basic module with an empty ' . $this->getExpectedFilenames())
+        $this->setName('make:file')
+            ->setDescription('Generate a ' . $this->getExpectedFilenames())
             ->addArgument('path', InputArgument::REQUIRED, 'The file path. For example "App/TestModule/TestSubModule"')
+            ->addArgument('filenames', InputArgument::REQUIRED | InputArgument::IS_ARRAY, $this->getExpectedFilenames())
             ->addOption('short-name', 's', InputOption::VALUE_NONE, 'Remove module prefix to the class name');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        /** @var list<string> $inputFileNames */
+        $inputFileNames = $input->getArgument('filenames');
+
+        $filenames = array_map(
+            fn (string $raw): string => $this->getFacade()->sanitizeFilename($raw),
+            $inputFileNames
+        );
+
         /** @var string $path */
         $path = $input->getArgument('path');
         $commandArguments = $this->getFacade()->parseArguments($path);
         $shortName = (bool)$input->getOption('short-name');
 
-        foreach (FilenameSanitizer::EXPECTED_FILENAMES as $filename) {
-            $fullPath = $this->getFacade()->generateFileContent(
+        foreach ($filenames as $filename) {
+            $absolutePath = $this->getFacade()->generateFileContent(
                 $commandArguments,
                 $filename,
                 $shortName
             );
-            $output->writeln("> Path '{$fullPath}' created successfully");
+            $output->writeln("> Path '{$absolutePath}' created successfully");
         }
-
-        $pieces = explode('/', $commandArguments->directory());
-        $moduleName = end($pieces);
-        $output->writeln("Module '{$moduleName}' created successfully");
 
         return self::SUCCESS;
     }
