+++ src/Console/Infrastructure/Template/Command/dependency-provider-maker.txt
