+++ src/Console/Infrastructure/Template/Command/facade-maker.txt
