+++ src/Console/Infrastructure/Template/Command/factory-maker.txt
